+++ CHANGELOG.md
@@ -10,6 +10,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.3] - 2019-02-01
+
+### Added
+- `Network` added, providing API methods for the blockstack network
+- `BlockstackSession.network` added, giving access to the blockstack network used in this session
+- `BlockstackSession.listFiles` iterates through the user's app files
+- `BlockstackSession.generateAndStoreTransitKey`, and `.redirectToSignInWithAuthRequest` to allow custom auth flow
+- `Scope.fromJSName` creates `Scope` from its javascript name as used in blockstack.js
+- `contentType` property in `PutFileOptions`
+
+
+### Changed
+- Use current blockstack.js version from feature/session-storage (18.2.1 with user session handling)
+- removed params from appConfig of `makeAuthRequest`
+
+
+
 ## [0.4.2] - 2018-12-08
 
 ### Changed